Fabo/simple node subscription (#4910)
* changelog

* poll for polkadot proposals to trigger events

* fix missing db

* cleanup

* standardize node subscriptions

* changelog

* lint

* overwrite old subscriptions
