+++ api/changes/fabo_simple-node-subscription
@@ -1,0 +1,1 @@
+[Code Improvements] Standardize the node subscription leveraging the source classes @faboweb
